requirements.txt
nextcord
python-dotenv
setuptools